Rename Smallest Index With Equal Value.py to 1365. HowManyNumbersAreSmallerThanTheCurrentNumber.py
Given the array nums, for each nums[i] find out how many numbers in the array are smaller than it. That is, for each nums[i] you have to count the number of valid j's such that j != i and nums[j] < nums[i].
Return the answer in an array.